pyproject.toml
[project]
name = "ml-peg"
version = "0.0.1"
description = "ML potential usability and performance guide"
authors = [
    { name = "Elliott Kasoar" },
    { name = "Joseph Hart" },
    { name = "Ilyes Batatia" },
    { name = "Alin M. Elena" },
    { name = "Gábor Csányi" },
]
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Natural Language :: English",
    "Development Status :: 3 - Alpha",
]
readme = "README.md"

dependencies = [
    "boto3<2,>=1.40.49",
    "dash>=3.1.1",
    "janus-core<1.0.0,>=0.8.2",
    "kaleido>=1.0.0",
    "mlipx<0.2,>=0.1.5",
    "scikit-learn>=1.7.1",
    "torch>=2.8.0",
    "typer<1.0.0,>=0.19.1",
]

[project.optional-dependencies]
chgnet = [
    "chgnet == 0.4.0",
]
d3 = [
    "torch-dftd==0.5.1",
]
dpa3 = [
    "deepmd-kit == 3.1.0",
]
grace = [
    "tensorpotential == 0.5.1; python_version < '3.13'",
]
mace = [
    "mace-torch==0.3.14",
]

mattersim = [
    "mattersim==1.2.0",
]
orb = [
    "orb-models == 0.5.5; sys_platform != 'win32' and python_version < '3.13'",
]

uma = [
    "fairchem-core == 2.10.0",
]

[project.scripts]
ml_peg = "ml_peg.cli.cli:app"

[project.urls]
Repository = "https://github.com/ddmms/ml-peg/"
Documentation = "https://ddmms.github.io/ml-peg/"

[dependency-groups]
dev = [
    "coverage[toml]<8.0.0,>=7.4.1",
    "ipykernel>=6.29.5",
    "pgtest<2.0.0,>=1.3.2",
    "pytest<9.0,>=8.0",
    "pytest-cov<5.0.0,>=4.1.0",
    "tox-uv<2.0,>=1.16.1",
    "wheel<1.0,>=0.42",
]

docs = [
    "furo<2025.0.0,>=2024.1.29",
    "jupyter>=1.1.1",
    "markupsafe",
    "nbsphinx>=0.9.6",
    "numpydoc<2.0.0,>=1.6.0",
    "sphinx<8.2.0,>=8.0.2",
    "sphinxcontrib-contentui<1.0.0,>=0.2.5",
    "sphinxcontrib-details-directive<1.0,>=0.1",
    "sphinx-autodoc-typehints<3.0.0,>=2.5.0",
    "sphinx-collapse>=0.1.3",
    "sphinx-copybutton<1.0.0,>=0.5.2",
]

pre-commit = [
    "pre-commit<4.0.0,>=3.6.0",
    "ruff<1.0.0,>=0.9.3",
]

prod = [
    "gunicorn>=23",
]

[build-system]
requires = ["uv_build"]
build-backend = "uv_build"

[tool.pytest.ini_options]
# Configuration for [pytest](https://docs.pytest.org)
python_files = "test_*.py"
addopts = '--cov-report xml'
pythonpath = ["."]

[tool.coverage.run]
# Configuration of [coverage.py](https://coverage.readthedocs.io)
# reporting which lines of your plugin are covered by tests
source=["ml_peg"]

[tool.ruff]
exclude = ["conf.py", "*ipynb"]
target-version = "py310"

[tool.ruff.lint]
# Ignore complexity
ignore = ["C901"]
select = [
    # flake8-bugbear
    "B",
    # pylint
    "C", "R",
    # pydocstyle
    "D",
    # pycodestyle
    "E", "W",
    # Pyflakes
    "F", "FA",
    # pyupgrade
    "I",
    # pep8-naming
    "N",
    # isort
    "UP",
]

[tool.ruff.lint.isort]
force-sort-within-sections = true
required-imports = ["from __future__ import annotations"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.pylint]
max-args = 10

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = false

[tool.numpydoc_validation]
# report on all checks, except the below
checks = [
    "all",
    "EX01",
    "SA01",
    "ES01",
    "PR04", # Ignore no type (types come from signature as via sphinx_autodoc_typehints)
]
# Don't report on objects that match any of these regex
exclude = [
    ".__weakref__$",
    ".__repr__$",
]

[tool.uv]
default-groups = [
    "dev",
    "docs",
    "pre-commit",
]

conflicts = [
    [
      { extra = "uma" },
      { extra = "mace" },
    ],
    [
      { extra = "uma" },
      { extra = "grace" },
    ],
    [
      { extra = "mattersim" },
      { extra = "mace" },
    ],
    [
      { extra = "mattersim" },
      { extra = "grace" },
    ],
]

constraint-dependencies = [
    "tensorflow>=2.16.1",
]

[tool.uv.build-backend]
module-root = ""
